v3 of story
Added a lot more flavor and cleared up some confusion hopefully. The big reveal is day six. At that point you learn that his daughter and the monster are different. I have tried to make it as explicit as possible without beating your head.

I made day seven pretty explicit about the threat. And then on day eight I put in the call to action. The uncle asks you to slay the beast and sacrifices himself to give you a head start. This also gives a 'realistic death' to the uncle (he died in a fire). If you want to include that at the start. Perhaps the start of the journal is burned. It should be pretty clear that the uncle died as well.

Finally I trimmed the rest of the story where I felt it could use it